# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreerEtablissementRequestDto" : {
        "properties" : {
          "adresseLigne" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "boitePostale" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "code" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "maxLength" : 320,
            "minLength" : 0,
            "type" : "string"
          },
          "emailAdministrateur" : {
            "maxLength" : 320,
            "minLength" : 0,
            "type" : "string"
          },
          "nom" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "nomCompletAdministrateur" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "quartier" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "sigle" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          },
          "siteWeb" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "telephone" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "typeEtablissement" : {
            "enum" : [
              "PRESCOLAIRE",
              "PRIMAIRE",
              "COLLEGE",
              "LYCEE",
              "COMPLEXE"
            ],
            "type" : "string"
          },
          "ville" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [
          "typeEtablissement"
        ],
        "type" : "object"
      },
...
      "EtablissementCreeDto" : {
        "properties" : {
          "etablissement" : {
            "$ref" : "#/components/schemas/EtablissementDto"
          },
          "motDePasseTemporaireAdmin" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "EtablissementDto" : {
        "properties" : {
          "actif" : {
            "type" : "boolean"
          },
          "adresseLigne" : {
            "type" : "string"
          },
          "boitePostale" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "dateCreation" : {
            "format" : "date-time",
            "type" : "string"
          },
          "dateModification" : {
            "format" : "date-time",
            "type" : "string"
          },
          "deviseCode" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "fuseauHoraire" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "langueDefaut" : {
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "paysCode" : {
            "type" : "string"
          },
          "quartier" : {
            "type" : "string"
          },
          "referentielInitialise" : {
            "type" : "boolean"
          },
          "sigle" : {
            "type" : "string"
          },
          "siteWeb" : {
            "type" : "string"
          },
          "telephone" : {
            "type" : "string"
          },
          "typeEtablissement" : {
            "enum" : [
              "PRESCOLAIRE",
              "PRIMAIRE",
              "COLLEGE",
              "LYCEE",
              "COMPLEXE"
            ],
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "EtablissementHistoriqueDto" : {
        "properties" : {
          "actif" : {
            "type" : "boolean"
          },
          "auteur" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "numeroRevision" : {
            "format" : "int64",
            "type" : "integer"
          },
          "typeRevision" : {
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "EtablissementResumeDto" : {
        "properties" : {
          "actif" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "nom" : {
            "type" : "string"
          },
          "nombreSites" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sigle" : {
            "type" : "string"
          },
          "typeEtablissement" : {
            "enum" : [
              "PRESCOLAIRE",
              "PRIMAIRE",
              "COLLEGE",
              "LYCEE",
              "COMPLEXE"
            ],
            "type" : "string"
          },
          "ville" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ModifierEtablissementRequestDto" : {
        "properties" : {
          "adresseLigne" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "boitePostale" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "deviseCode" : {
            "maxLength" : 3,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "maxLength" : 320,
            "minLength" : 0,
            "type" : "string"
          },
          "fuseauHoraire" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "langueDefaut" : {
            "maxLength" : 10,
            "minLength" : 0,
            "type" : "string"
          },
          "nom" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "quartier" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "sigle" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          },
          "siteWeb" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "telephone" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "typeEtablissement" : {
            "enum" : [
              "PRESCOLAIRE",
              "PRIMAIRE",
              "COLLEGE",
              "LYCEE",
              "COMPLEXE"
            ],
            "type" : "string"
          },
          "ville" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [
          "typeEtablissement"
        ],
        "type" : "object"
      },
...
      "PageReponseEtablissementResumeDto" : {
        "properties" : {
          "contenu" : {
            "items" : {
              "$ref" : "#/components/schemas/EtablissementResumeDto"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "taille" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "in" : "header",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API de gestion scolaire Edukeys, éditée par Nova Digital (Togo).",
    "title" : "Edukeys — API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/etablissements" : {
      "get" : {
        "operationId" : "listerEtablissements",
        "parameters" : [
          {
            "in" : "query",
            "name" : "page",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          },
          {
            "in" : "query",
            "name" : "size",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageReponseEtablissementResumeDto"
                }
              }
            },
            "description" : "Page d'établissements"
          }
        },
        "summary" : "Liste paginée de tous les établissements de la plateforme (opération SUPER_ADMIN)",
        "tags" : [
          "Établissements"
        ]
      },
      "post" : {
        "operationId" : "creerEtablissement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreerEtablissementRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EtablissementCreeDto"
                }
              }
            },
            "description" : "Établissement créé, avec le mot de passe temporaire de son premier administrateur (retourné une seule fois)"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EtablissementCreeDto"
                }
              }
            },
            "description" : "Code ou email déjà porté par un établissement actif"
          }
        },
        "summary" : "Crée un établissement (identité, coordonnées, site principal, premier compte ADMIN et référentiel pédagogique initialisés en une seule transaction)",
        "tags" : [
          "Établissements"
        ]
      }
    },
...
    "/api/v1/etablissements/{id}" : {
      "get" : {
        "operationId" : "obtenirEtablissement",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EtablissementDto"
                }
              }
            },
            "description" : "Établissement trouvé"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EtablissementDto"
                }
              }
            },
            "description" : "Établissement introuvable"
          }
        },
        "summary" : "Détail d'un établissement",
        "tags" : [
          "Établissements"
        ]
      },
      "put" : {
        "operationId" : "modifierEtablissement",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ModifierEtablissementRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EtablissementDto"
                }
              }
            },
            "description" : "Établissement modifié"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EtablissementDto"
                }
              }
            },
            "description" : "Établissement introuvable"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EtablissementDto"
                }
              }
            },
            "description" : "Email déjà porté par un autre établissement actif"
          }
        },
        "summary" : "Modifie l'identité et les coordonnées d'un établissement (le code est immuable)",
        "tags" : [
          "Établissements"
        ]
      }
    },
    "/api/v1/etablissements/{id}/desactivation" : {
      "post" : {
        "operationId" : "desactiverEtablissement",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Établissement désactivé"
          }
        },
        "summary" : "Désactive un établissement (et, en cascade logique, ses sites et son logo)",
        "tags" : [
          "Établissements"
        ]
      }
    },
    "/api/v1/etablissements/{id}/historique" : {
      "get" : {
        "operationId" : "listerHistoriqueEtablissement",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EtablissementHistoriqueDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Révisions de l'établissement"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EtablissementHistoriqueDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Établissement introuvable"
          }
        },
        "summary" : "Historique des révisions d'un établissement",
        "tags" : [
          "Établissements"
        ]
      }
    },
    "/api/v1/etablissements/{id}/reactivation" : {
      "post" : {
        "operationId" : "reactiverEtablissement",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "Établissement réactivé"
          },
          "409" : {
            "description" : "Code ou email repris par un autre établissement actif entre-temps"
          }
        },
        "summary" : "Réactive un établissement, si son code et son email restent disponibles",
        "tags" : [
          "Établissements"
        ]
      }
    },
...
  },
  "security" : [
    {
      "bearerAuth" : [ ]
    }
  ],
  "servers" : [
    {
      "description" : "Generated server url",
      "url" : "http://localhost"
    }
  ]
}